Renamed the modules by their category as the problems solved will not only be from a book.
