+++ src/lib.rs
@@ -1,8 +1,4 @@
-mod ex_01_01;
-mod ex_01_02;
-mod ex_01_03;
-mod ex_01_04;
-mod ex_01_05;
-mod ex_02;
-mod ex_03;
+mod string;
+mod linked_list;
+mod search;
 mod data_structures;
