+++ src/linked_list/ex_02_01.rs
