+++ src/string/ex_01_01.rs
