+++ src/string/ex_01_02.rs
