+++ src/string/ex_01_05.rs
