+++ src/string/mod.rs
@@ -1,0 +1,5 @@
+pub mod ex_01_01;
+pub mod ex_01_02;
+pub mod ex_01_03;
+pub mod ex_01_04;
+pub mod ex_01_05;
